requirements.txt
streamlit
python-dotenv
google-generativeai
pillow
pandas
numpy
matplotlib
plotly
scikit-learn
seaborn
